search: single matches event
This updates the streaming backend to only send a single matches event
type. This simplifies backend code, and will also simplify frontend
code (which currently combines them).

This commits is not functional yet since the backend needs to tag each
match with the type, and the frontend needs to be updated to accept the
new event type.